JusticeAI.API/requirements.txt
fastapi
python-dotenv
opencensus
opencensus-ext-azure
gunicorn
uvicorn
tiktoken
unstructured
azure-identity
python-multipart
langchain-community
langchain-openai
langchain
azure-search-documents
azure-search
psutil
retry
requests
azure-cognitiveservices-vision-computervision
